update
correct  some  errors in MDAC_program.c

+++ 02-MCAL/08-MDAC/MDAC_program.c
@@ -17,27 +17,26 @@
 #include "MDAC_config.h"
 #include "MDAC_PRIVATE.h"
 
-volatile u8 sample_counter=0;
-
+volatile u32 sample_counter=0;
 
 
 static void MDAC_voidCALLBACK(void)
 {
-    MGPIO_voidSetPinValue(MDAC_PortBit0,MDAC_PinBit0,GET_BIT(song_raw[sample_counter],0));
-    MGPIO_voidSetPinValue(MDAC_PortBit1,MDAC_PinBit1,GET_BIT(song_raw[sample_counter],1));
-    MGPIO_voidSetPinValue(MDAC_PortBit2,MDAC_PinBit2,GET_BIT(song_raw[sample_counter],2));
-    MGPIO_voidSetPinValue(MDAC_PortBit3,MDAC_PinBit3,GET_BIT(song_raw[sample_counter],3));
-    MGPIO_voidSetPinValue(MDAC_PortBit4,MDAC_PinBit4,GET_BIT(song_raw[sample_counter],4));
-    MGPIO_voidSetPinValue(MDAC_PortBit5,MDAC_PinBit5,GET_BIT(song_raw[sample_counter],5));
-    MGPIO_voidSetPinValue(MDAC_PortBit6,MDAC_PinBit6,GET_BIT(song_raw[sample_counter],6));
-    MGPIO_voidSetPinValue(MDAC_PortBit7,MDAC_PinBit7,GET_BIT(song_raw[sample_counter],7));
+    MGPIO_voidSetPinValue(MDAC_PortBit0,MDAC_PinBit0,GET_BIT((u8)song_raw[sample_counter],0));
+    MGPIO_voidSetPinValue(MDAC_PortBit1,MDAC_PinBit1,GET_BIT((u8)song_raw[sample_counter],1));
+    MGPIO_voidSetPinValue(MDAC_PortBit2,MDAC_PinBit2,GET_BIT((u8)song_raw[sample_counter],2));
+    MGPIO_voidSetPinValue(MDAC_PortBit3,MDAC_PinBit3,GET_BIT((u8)song_raw[sample_counter],3));
+    MGPIO_voidSetPinValue(MDAC_PortBit4,MDAC_PinBit4,GET_BIT((u8)song_raw[sample_counter],4));
+    MGPIO_voidSetPinValue(MDAC_PortBit5,MDAC_PinBit5,GET_BIT((u8)song_raw[sample_counter],5));
+    MGPIO_voidSetPinValue(MDAC_PortBit6,MDAC_PinBit6,GET_BIT((u8)song_raw[sample_counter],6));
+    MGPIO_voidSetPinValue(MDAC_PortBit7,MDAC_PinBit7,GET_BIT((u8)song_raw[sample_counter],7));
     sample_counter++;
 
     if(sample_counter==song_length)
     {
       sample_counter=0;
     }
-
+    
 
 
 }
